@arxanas 2023-07-12
I extracted the scarier stuff into https://github.com/martinvonz/jj/pull/1858 (which is failing tests), but even without those commits, there is a performance regression in this PR when not using Watchman:

```
# main, no-fsmonitor
$ hyperfine --prepare 'cargo build --release --features watchman' --warmup 3 './target/release/jj -R ../nixpkgs status'

Benchmark 1: ./target/release/jj -R ../nixpkgs status
  Time (mean ± σ):      1.590 s ±  0.027 s    [User: 0.511 s, System: 1.039 s]
  Range (min … max):    1.560 s …  1.642 s    10 runs

# pr, no-fsmonitor
$ hyperfine --prepare 'cargo build --release --features watchman' --warmup 3 './target/release/jj -R ../nixpkgs status'

Benchmark 1: ./target/release/jj -R ../nixpkgs status
  Time (mean ± σ):      2.827 s ±  0.034 s    [User: 1.465 s, System: 1.320 s]
  Range (min … max):    2.778 s …  2.895 s    10 runs
```

So I'll need to investigate.