@arxanas 2023-07-12
Bisected to 7e27acc "refactor(working_copy): return updated new tree value from `update_file_state` (the first commit in the stack 😬). Possibly it has to do with superfluous tree traversals, since I didn't do anything like hold a reference to the current nested `Tree`.