@martinvonz 2023-07-15
> Bisected to [7e27acc](https://github.com/martinvonz/jj/commit/7e27acc149a338ad8c9bd247c21a3899dd44f5c9) "refactor(working_copy): return updated new tree value from `update_file_state` (the first commit in the stack 😬). Possibly it has to do with superfluous tree traversals, since I didn't do anything like hold a reference to the current nested `Tree`.

I suspect it's because the new code read the current `TreeValue` and updates it in the `TreeBuilder` for every path even if it didn't change.

I feel like we should be able to visit only the paths that Watchman tells us and update the `TreeBuilder` for those paths. We would then bypass every use of `TreeState::file_states` when using Watchman. Hmm, since we store more than that the file states in the same proto, we would also have to split up the file so we can avoid reading and writing the file states when using Watchman. Thoughts?